BmpSplitter と BmpMuxer の実装を更新して自前処理で分割と結合を行うように変更
こうする事でファイルに出力されるバイト配列の内容の細かいコントロールが可能になる
