+++ tools/ToolBarTools/ToolBarImageSplitter/BmpSplitter.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.IO;
-using System.Drawing;
 using System.Diagnostics;
+using ToolBarImageCommon;
 
 namespace ToolBarImageSplitter
 {
@@ -18,9 +18,10 @@ namespace ToolBarImageSplitter
                 Directory.CreateDirectory(outDir);
             }
 
-            var bmp = (Bitmap)Image.FromFile(fileName);
-            var width = bmp.Size.Width;
-            var height = bmp.Size.Height;
+            var bmp = Bmp.FromFile(fileName);
+            var width = bmp.bmih.biWidth;
+            var height = Math.Abs(bmp.bmih.biHeight);
+            // 縦横ピクセル数同じ限定
             var sx = width / countPerLine;
             var sy = sx;
 
@@ -30,7 +31,6 @@ namespace ToolBarImageSplitter
                 for (var x = 0; x < width; x += sx)
                 {
                     var outfile = string.Empty;
-                    
                     if (x + sx == width)
                     {
 	                    // 右端の画像はインデックス用
@@ -41,20 +41,9 @@ namespace ToolBarImageSplitter
                         index++;
                         outfile = Path.Combine(outDir, index.ToString() + ".bmp");
                     }
-
-                    var cloneRect = new RectangleF(x, y, sx, sy);
-                    using (var cloneBitmap = bmp.Clone(cloneRect, bmp.PixelFormat))
-                    {
-                        cloneBitmap.Palette = bmp.Palette;
-                        cloneBitmap.Save(outfile, System.Drawing.Imaging.ImageFormat.Bmp);
-                        Console.WriteLine("wrote {0} x={1} y={2}", outfile, x, y);
-                    }
-#if DEBUG
-                    using (var bmpDebug = (Bitmap)Image.FromFile(outfile))
-                    {
-                        Debug.Assert(bmpDebug.Palette.Entries.SequenceEqual(bmp.Palette.Entries));
-                    }
-#endif
+                    var copiedBitmap = bmp.Copy(x, y, sx, sy);
+                    copiedBitmap.ToFile(outfile);
+                    Console.WriteLine("wrote {0} x={1} y={2}", outfile, x, y);
                 }
             }
         }
