+++ tools/ToolBarTools/ToolBarTools.sln
@@ -4,8 +4,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 15.0.28307.960
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ToolBarImageSplitter", "ToolBarImageSplitter\ToolBarImageSplitter.csproj", "{4F23CE26-8A95-44AD-BE88-45DFE08FE0F7}"
+	ProjectSection(ProjectDependencies) = postProject
+		{84E29976-DFDC-4661-9316-F434B42F75C7} = {84E29976-DFDC-4661-9316-F434B42F75C7}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ToolBarImageMuxer", "ToolBarImageMuxer\ToolBarImageMuxer.csproj", "{ABF37460-F080-4F1A-9F7A-F166154072FF}"
+	ProjectSection(ProjectDependencies) = postProject
+		{84E29976-DFDC-4661-9316-F434B42F75C7} = {84E29976-DFDC-4661-9316-F434B42F75C7}
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ToolBarImageCommon", "ToolBarImageCommon\ToolBarImageCommon.csproj", "{84E29976-DFDC-4661-9316-F434B42F75C7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +29,10 @@ Global
 		{ABF37460-F080-4F1A-9F7A-F166154072FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ABF37460-F080-4F1A-9F7A-F166154072FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ABF37460-F080-4F1A-9F7A-F166154072FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{84E29976-DFDC-4661-9316-F434B42F75C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{84E29976-DFDC-4661-9316-F434B42F75C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{84E29976-DFDC-4661-9316-F434B42F75C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{84E29976-DFDC-4661-9316-F434B42F75C7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
